backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.97.0

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.21.1
httpx==0.26.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Added from the code block
requests==2.31.0 